(1)) Comments added to eapl classes. (2) Identifier db table added (additionally entity + repository). (3) Db repositories fixed (uuid -> Integer), (4) mvn spotless:apply

+++ erzwingungshaft-eai/src/main/java/de/muenchen/eh/infrastructure/db/repository/ClaimDocumentRepository.java
@@ -6,11 +6,9 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClaimDocumentRepository extends CrudRepository<ClaimDocument, Long> {
+public interface ClaimDocumentRepository extends CrudRepository<ClaimDocument, Integer> {
 
     List<ClaimDocument> findByClaimImportIdOrderByDocumentType(Integer claimImportId);
 
     List<ClaimDocument> findByDocumentType(String type);
 }
-
-
